removed profile endpoints
